Main Project Section was Added
In this commit, we have created a section for representing Main Projects. This Section contains an animated Header and slider to show the projects, and also a scale and rotate effect on the image when hovering.

Now the user can see Main Projects.
New Libraries: atropos, swiper

index.css now contains all the common CSS of the project.
Created a useActiveScroll hook to detect if a component is in the viewport or not.