Added Anchor Tag to the Project Image
User now can directly click on the Image to open the live demo, if live
demo is not available then code else no click feature is available.

The Project Component was modified by adding a anchor tag to the image

Note:
Debugging of this feature is remaining, some unknown issues can occur.